Added commands and reworked ui and village relations
added more commands and fleshed out the print method,(still need to
figure out text color). reworked the relationship between the village
object and the UI object. when a UI is created it requires a village to
be passed in. Might need to be the other way around later to support
multi-village setups.

started to change all the system.out.println's to my custom print method
for the console