solved the text color issue and added the findVillager method
solved the text color, the console textpane was disabled rather than set
to not editable. when an element is disabled it becomes black and white
thus showing the wrong color.

added a findVillager method to locate existing villagers in the village.
they are held in an arraylist and the ui searches through that list for
the object matching the name given.