reworked the mayor class and added commands
reworked the mayor class to extend the villager class, this inherits the
methods such as eat sleep etc. instead of requiring an already existing
villager to create a mayor, the mayor class creates a new Villager
similar to how the dog or bird class creates a new Villager.

add more commands to list the Villagers in the villagerList. Thought
about making a list for all other objects to find things like houses but
couldn't think of a way to iterate through it efficiently.